Fix: condition for ranged queries with query operators now work. Also simplified query conversion of numerical values.
